hospital-his-ml/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
pymongo>=4.6.0
python-dotenv>=1.0.0
joblib>=1.3.0
prophet>=1.1.5
statsmodels>=0.14.1
matplotlib>=3.8.0